implementa classe pedido e seus metodos do Crud e validações necessarias para compra

+++ src/main/java/com/prata/venda/product/application/api/ProductAlteracaoRequest.java
@@ -2,10 +2,11 @@ package com.prata.venda.product.application.api;
 
 import com.prata.venda.product.domain.Promotion;
 import lombok.Value;
+import java.math.BigDecimal;
 
 @Value
 public class ProductAlteracaoRequest {
     String name;
-    double price;
+    BigDecimal price;
     Promotion promotion;
 }
